Add files via upload
Hi there!😎 I am writing to express my interest in the role at African Leadership Group. As a 32-year-old Ruby on Rails developer, I am confident that I have the skills and experience needed to make a valuable contribution to your team. 

I am also a friendly and outgoing individual, who is always eager to collaborate with others and find new ways to bring my ideas to life. I believe that my positive attitude, combined with my technical expertise, would make me a great fit for the role.

Reading about the vision and mission of African Leadership Group and ALX has inspired me and I am excited to be a part of the team that is working towards developing the skills for the fourth industrial revolution and helping Africa compete in the digital world. The fact that ALX has signed an agreement to acquire Silicon Valley based Holberton Inc. is evidence of its commitment to providing the best technology education and talent to the world.

I am especially interested in the role as it will allow me to be directly involved in the student's journey, assist the team in tracking copyright violations, and help design and implement automations. My great communication skills, ability to manage projects independently, and experience with tools like Slack, Google Drive, Git and GitHub would allow me to thrive in this role.

My passion for humanity, desire to create lasting impact, and ability to think big, combined with my disciplined approach, make me a strong candidate for this role. I also enjoy educating others